+++ Nef_3/doc/Nef_3/CGAL/boost/graph/convert_nef_polyhedron_to_polygon_mesh.h
@@ -6,7 +6,12 @@ namespace CGAL {
 /// (but faces with more than one connected component of the boundary).
 /// The polygon mesh can be triangulated by setting `triangulate_all_faces` to `true` or by calling the function `triangulate_faces()`.
 /// @tparam Nef_polyhedron an object of type `Nef_polyhedron_3`.
-/// @tparam Polygon_mesh a model of `MutableFaceGraph`.
+/// @tparam Polygon_mesh a model of `MutableFaceGraph` with an internal property map for `CGAL::vertex_point_t`.
+///
+/// The points from `nef` to `pm` are converted using
+/// `CGAL::Cartesian_converter<NefKernel, TargetKernel>`.
+/// `NefKernel` and `TargetKernel` are deduced using `CGAL::Kernel_traits`
+/// from the point type of `nef` and the value type of the vertex_point_map of `tm`.
 /// 
 /// @param nef the input.
 /// @param pm the output.
@@ -21,18 +26,25 @@ namespace CGAL {
   /// \ingroup PkgNef3IOFunctions
   /// Converts an objet of type `Nef_polyhedron_3` into a polygon soup.
   /// The polygons can be triangulated by setting `triangulate_all_faces` to `true`.
-  /// @tparam Output_kernel the Kernel from which the point type of points come from.
   /// @tparam Nef_polyhedron an object of type `Nef_polyhedron_3`.
   /// @tparam PointRange a model of the concepts `RandomAccessContainer` and 
   /// `BackInsertionSequence` whose `value_type` is the point type
-  /// @tparam PolygonRange a model of the concept `RandomAccessContainer` whose 
-  /// `value_type` is a model of the concept `RandomAccessContainer` whose `value_type` is `std::size_t`.
+  /// @tparam PolygonRange a model of the concepts `RandomAccessContainer` and 
+  /// `BackInsertionSequence` whose `value_type` is a model of the concepts 
+  /// `RandomAccessContainer` and `BackInsertionSequence` whose 
+  /// `value_type` is `std::size_t`.
+  /// It must 
+  /// 
+  /// The points from `nef` to `points` are converted using
+  /// `CGAL::Cartesian_converter<NefKernel, OutputKernel>`.
+  /// `NefKernel` and `OutputKernel` are deduced using `CGAL::Kernel_traits`
+  /// from the point types.
   /// 
   /// @param nef the input.
   /// @param points the output points of the soup
   /// @param polygons the output polygons of the soup. 
   /// @param triangulate_all_faces indicates whether all polygons must be triangulated.
-  template <class Output_kernel, class Nef_polyhedron, typename PolygonRange, typename PointRange>
+  template <class Nef_polyhedron, typename PolygonRange, typename PointRange>
   void convert_nef_polyhedron_to_polygon_soup(const Nef_polyhedron& nef,
                                                     PointRange& points,
                                                     PolygonRange& polygons,