+++ Installation/CHANGES.md
@@ -6,7 +6,7 @@ Release 4.15
 Release date: September 2019
 
 ###3D Boolean Operations on Nef Polyhedra
--   Added a function to convert a Nef_polyhedron to a polygon soup:
+-   Added a function to convert a Nef_polyhedron_3 to a polygon soup:
     - `CGAL::convert_nef_to_polygon_soup()`
 
 Release 4.14